+++ lib/stattleship/params/total_team_stat_params.rb
@@ -1,14 +1,14 @@
 module Stattleship
   module Params
-    class TotalTeamStatParams < Stattleship::Params::QueryParams
-      attr_accessor :interval_type, :team_id, :stat
+    class TotalTeamStatParams < Stattleship::Params::GameTimeParams
+      attr_accessor :interval_type, :team_id, :stat, :game_id
       attr_reader :type
 
       private
 
       def params
         {
-          'interval_type' => interval_type,
+          'game_id' => game_id,
           'team_id' => team_id,
           'stat' => stat,
           'type' => type,